+++ lib/factory_sloth/create_call_finder.rb
@@ -7,29 +7,44 @@ class FactorySloth::CreateCallFinder < Ripper
     new(code).tap(&:parse).locations
   end
 
-  def initialize(...)
+  def initialize(code, ...)
     super
+    @code = code
+    @disabled = false
     @locations = []
   end
   private_class_method :new
+
+  def store_location(loc)
+    @locations << loc if loc.instance_of?(Array) && !@disabled
+  end
 
   def on_ident(name, *)
     [lineno, column] if %w[create create_list create_pair].include?(name)
   end
 
   def on_call(mod, _, loc, *)
-    @locations << loc if loc.instance_of?(Array) && mod == 'FactoryBot'
+    store_location(loc) if mod == 'FactoryBot'
   end
 
   def on_command_call(mod, _, loc, *)
-    @locations << loc if loc.instance_of?(Array) && mod == 'FactoryBot'
+    store_location(loc) if mod == 'FactoryBot'
+  end
+
+  def on_comment(text, *)
+    return unless /sloth:(?<directive>disable|enable)/ =~ text
+
+    directive == 'disable' &&
+      @locations.reject! { |loc| loc.first == lineno } ||
+      (@lines ||= @code.lines)[lineno - 1].match?(/^\s*#/) &&
+      (@disabled = directive != 'enable')
   end
 
   def on_fcall(loc, *)
-    @locations << loc if loc.instance_of?(Array)
+    store_location(loc)
   end
 
   def on_vcall(loc, *)
-    @locations << loc if loc.instance_of?(Array)
+    store_location(loc)
   end
 end